+++ src/main/resources/db/migration/20171120155419223__change_supervisory_node_columns.sql
@@ -1,0 +1,2 @@
+ALTER TABLE supervisory_nodes ALTER COLUMN facilityid DROP NOT NULL;
+ALTER TABLE supervisory_nodes ALTER COLUMN name SET NOT NULL;