+++ src/test/java/org/openlmis/referencedata/domain/SupervisoryNodeTest.java
@@ -40,7 +40,7 @@ public class SupervisoryNodeTest {
   public void setUp() {
     program = new Program("P1");
     facility2 = new Facility("C2");
-    supervisoryNode1 = SupervisoryNode.newSupervisoryNode("SN1", new Facility("C1"));
+    supervisoryNode1 = SupervisoryNode.newSupervisoryNode("node", "SN1", new Facility("C1"));
     requisitionGroup1 = new RequisitionGroup("RG1", "RGN1", supervisoryNode1);
     requisitionGroup1.setMemberFacilities(Sets.newHashSet(facility2, new Facility("C3")));
     addSupportedPrograms(requisitionGroup1);
@@ -66,7 +66,7 @@ public class SupervisoryNodeTest {
   public void shouldGetAllIndirectSupervisedFacilities() {
     //given
     SupervisoryNode supervisoryNode2 =
-        SupervisoryNode.newSupervisoryNode("SN2", new Facility("C4"));
+        SupervisoryNode.newSupervisoryNode("node", "SN2", new Facility("C4"));
     RequisitionGroup requisitionGroup2 = new RequisitionGroup("RG2", "RGN2", supervisoryNode2);
     requisitionGroup2.setMemberFacilities(Sets.newHashSet(new Facility("C5")));
     addSupportedPrograms(requisitionGroup2);